Fixed Map not being able to have new values assigned to it.
By storing shared_ptr of objects instead of Pointers of objects, those can be constructed as empty when making space for a new key's value and then set to their correct shared_ptr to the new value.
Pointers can't be null and therefore were not allowing this to happen inside the map template.
operator[] was replaced by valueForKey() and setValueForKey() in order to be consistent. Otherwise operator[] would have been used for reading and setValurForKey() for writing/creating.