Simplified the internal pointer for all Objects.
internal is now a regular raw pointer. Dereferencing it is much faster than doing it thru a Pointer each time. As a resource, internal is stored in the base Object class with a Pointer to the internal object. Note that internally this is using shared_ptr which guarantees that the correct destructor is called for the internal object when the base object is destroyed.