Got rid of the destroy() methods.
Objects are not deleted by any specific methods now, they are just removed from their parents and let to be deleted by the last Pointer that still refers to them. TrackEntry and Crates can be removed from their parent either with a call in the parent or a call in the object. The result is the same.