Added a macro to generate a default operator equal for Objects.
Operator equal with its own class is not generated by default. Because of the operator equal which generated to test any class against an Object (allowing for polymorphic tests from the base class) there is no way to fall back down to a compiler generated operator equal. Instead, if the class doesn't want to implement its own operator equal then it can use this macro to get a default one which tests if the object is EXACTLY the same as the other one (addresses are equal).