Refactored TrackFiles so that they cache their data and work around a TagLib design 'feature'.
TagLib doesn't save its files when it says it does. In order to make sure save means save, we cache the data and only open the file once to parse and then once to save. When we save, we use TagLib's fake save method and then destroy the TagLib file to make sure save really happens.