Unique objects now have a deleted default copy constructor.
This is to prevent users from being able to make a non-const reference from a const reference, defeating the purpose of the const keyword.

In order for unique objects to be used in const containers, containers need to alwasy return references so that nothing needs to be constructed (a const unique object would not be able to be constructed) and objects must be passed to containers by value because otherqise a const reference to a unique object would not allow the containter to construct a storage object for the value.

Also in this commit, internal objects can now have pure virtual base classes.