Added classNameIs() method and fixed inheritance of className() and classHash().
Static versions of className() and classHash() can return the result from the current class but the instance versions need to call the internal object in order to return any inherited results instead.

i.e. calling className() on a base class need to return the class name of the actual object, not the base class object.